Segregate large file StreamReader test
